server/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
jinja2
